fix:update limit type

+++ src/main/java/com/sprint/findex_team6/service/IndexValService.java
@@ -238,7 +238,7 @@ public class IndexValService {
   }
 
   //특정 지수 성과 순위 조회
-  public List<RankedIndexPerformanceDto> getIndexPerformanceRank(String periodType, Long indexInfoId, Long limit) {
+  public List<RankedIndexPerformanceDto> getIndexPerformanceRank(String periodType, Long indexInfoId, int limit) {
     LocalDate beforeDate = calculateStartDate(periodType);
     LocalDate today = LocalDate.now();
 
@@ -275,7 +275,7 @@ public class IndexValService {
         .collect(Collectors.toList());
   }
   //indexInfoId 받지 않을 때 로직
-  public List<RankedIndexPerformanceDto> getIndexPerformanceRank(String periodType, Long limit) {
+  public List<RankedIndexPerformanceDto> getIndexPerformanceRank(String periodType, int limit) {
     LocalDate beforeDate = calculateStartDate(periodType);
     LocalDate today = LocalDate.now();
